turn on log switch

+++ VTMagicDemo/Classes/Home/VTGridViewController.m
@@ -84,7 +84,7 @@ static NSString *reuseIdentifier = @"grid.reuse.identifier";
 {
     VTGridViewCell *cell = [collectionView dequeueReusableCellWithReuseIdentifier:reuseIdentifier forIndexPath:indexPath];
     cell.backgroundColor = [UIColor whiteColor];
-    NSString *imageName = [NSString stringWithFormat:@"image_%ld", indexPath.row%13];
+    NSString *imageName = [NSString stringWithFormat:@"image_%ld", (long)indexPath.row%13];
     [cell.imageView setImage:[UIImage imageNamed:imageName]];
     cell.commentLabel.text = [NSString stringWithFormat:@"%d人出游", arc4random_uniform(9999)];
     cell.titleLabel.text = @"景点介绍，景点介绍，景点介绍。。";
